arduino-uno/live_graph/requirements.txt
pyserial
matplotlib
numpy
